Rename ezasm components to rezasm
