+++ .gitignore
@@ -1,3 +1,3 @@
 /target
 /.idea
-/ezasm-core/target/
+/rezasm-core/target/
