+++ Cargo.toml
@@ -4,8 +4,8 @@ version = "0.3.0"
 edition = "2021"
 
 [dependencies]
-ezasm-core = { path = "ezasm-core" }
-ezasm-macro = { path = "ezasm-macro" }
+rezasm-core = { path = "rezasm-core" }
+rezasm-macro = { path = "rezasm-macro" }
 
 lazy_static = "1.4.0"
 regex = "1.8.1"
@@ -14,4 +14,4 @@ bytebuffer = "2.1.0"
 clap = { version = "4.4.1", features = ["derive"] }
 
 [workspace]
-members = ["ezasm-core", "ezasm-macro"]
+members = ["rezasm-core", "rezasm-macro"]
