+++ rezasm-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ezasm-core"
+name = "rezasm-core"
 version = "0.1.0"
 edition = "2021"
