+++ rezasm-core/src/instructions/instruction_registry.rs
