+++ rezasm-core/src/simulation/memory.rs
