+++ rezasm-core/src/simulation/mod.rs
