+++ rezasm-core/src/simulation/registry.rs
