+++ rezasm-core/src/util/error.rs
