+++ rezasm-core/src/util/mod.rs
