+++ rezasm-core/src/util/word_size.rs
