+++ src/instructions/implementation/arithmetic_instructions.rs
@@ -1,6 +1,6 @@
-use ezasm_core::instructions::instruction_field::InstructionField;
-use ezasm_core::instructions::instruction_registry::register_instruction;
-use ezasm_macro::instruction;
+use rezasm_core::instructions::instruction_field::InstructionField;
+use rezasm_core::instructions::instruction_registry::register_instruction;
+use rezasm_macro::instruction;
 use lazy_static::lazy_static;
 
 lazy_static! {
